+++ src/helpers/productHelper.mjs
@@ -1,0 +1,5 @@
+import * as productService from '../services/productService.mjs';
+
+export const getAllProducts = () => {
+    return productService.fetchAllProducts();
+};
