+++ src/helpers/responseHelper.mjs
@@ -1,0 +1,6 @@
+export const getErrorResponse = (message) => {
+    return {
+        error: true,
+        message: message
+    };
+};
