+++ src/services/productService.mjs
@@ -1,0 +1,7 @@
+const mockProducts = [
+    { id: 123, name: 'Chicken Breast', price: 12.99 }
+];
+
+export const fetchAllProducts = () => {
+    return mockProducts;
+};
